requirements.txt
boto3==1.35.7
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Dropzone==2.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
ipykernel==6.29.5
matplotlib==3.9.2
nbformat==5.10.4
pandas==2.2.2
pillow==10.4.0
plotly==5.23.0
polars==1.6.0
requests==2.32.3
scipy==1.14.1
seaborn==0.13.2
snowflake-connector-python==3.12.1
SQLAlchemy==2.0.32
WTForms==3.1.2
email_validator==2.2.0
PyJWT==2.9.0
Flask-Dropzone==2.0.0
#itsdangerous==2.0.1 pip install --force-reinstall itsdangerous==2.0.1 swapped for PyJWT
redis==5.0.8
pyarrow==17.0.0
duckdb==1.0.0
#Faker==28.4.1
#fastparquet==2024.5.0
#dask==2024.8.2
simplejson==3.19.3
Flask-Limiter==3.8.0